requirements.txt
ib_insync>=0.9.86
pandas>=2.0.0
pandas_market_calendars>=4.3.1
pyyaml>=6.0.1
python-dateutil>=2.8.2
pytz>=2023.3 